requirements.txt
Django~=3.1.6
Pillow~=8.1.0
btc-template-tables~=0.4.2
django-filter~=2.4.0